feature: Testes de PedidoService implementados.

+++ src/main/java/com/fiap/tc/ms/gestao_pedidos/dto/response/PedidoPaginadoResponse.java
@@ -2,9 +2,11 @@ package com.fiap.tc.ms.gestao_pedidos.dto.response;
 
 import com.fiap.tc.ms.gestao_pedidos.model.enums.StatusPedido;
 
+import java.math.BigDecimal;
+
 public record PedidoPaginadoResponse(
     Long pedidoId,
     StatusPedido statusPedido,
-    double valorTotal
+    BigDecimal valorTotal
 ) {
 }
